+++ include/objc-helpers/BlockUtil.h
@@ -9,6 +9,10 @@
 #ifndef HEADER_BLOCK_UTIL_INCLUDED
 #define HEADER_BLOCK_UTIL_INCLUDED
 
+#if !__cpp_concepts
+    #error This header requires C++20 mode or above with concepts support
+#endif
+
 #include <concepts>
 #include <type_traits>
 #include <cstdlib>